service registeration simplify
