--- Dayana/Server/Api/Extensions/DependencyInjection/DatabaseInjection.cs
@@ -1,16 +1,0 @@
-﻿using Dayana.Shared.Persistence.EntityFrameWorkObjects;
-using Microsoft.EntityFrameworkCore;
-
-namespace Dayana.Server.Api.Extensions.DependencyInjection;
-
-public static class DatabaseInjection
-{
-    public static IServiceCollection AddConfiguredDatabase(this IServiceCollection services, IConfiguration configuration)
-    {
-        services.AddDbContext<AppDbContext>(options =>
-            options.UseSqlServer(configuration.GetConnectionString("ServerDbConnection"))
-            .EnableDetailedErrors());
-
-        return services;
-    }
-}
