--- Dayana/Server/Api/Extensions/DependencyInjection/HealthCheckInjection.cs
@@ -1,14 +1,0 @@
-﻿using Dayana.Shared.Basic.MethodsAndObjects.HealthChecks;
-
-namespace Dayana.Server.Api.Extensions.DependencyInjection;
-
-public static class HealthCheckInjection
-{
-    public static IServiceCollection AddConfiguredHealthChecks(this IServiceCollection services)
-    {
-        services.AddHealthChecks()
-            .AddCheck<GeneralHealthCheck>("general-check");
-
-        return services;
-    }
-}
